+++ app/src/main/java/fr/jaetan/jmedia/app/settings/SettingsView.kt
@@ -1,14 +1,27 @@
 package fr.jaetan.jmedia.app.settings
 
-import androidx.compose.material3.Text
+import androidx.compose.foundation.layout.Box
 import androidx.compose.runtime.Composable
-import fr.jaetan.jmedia.R
-import fr.jaetan.jmedia.extensions.localized
+import fr.jaetan.jmedia.app.settings.views.ContentView
+import fr.jaetan.jmedia.app.settings.views.RemoveDataDialog
+import fr.jaetan.jmedia.app.settings.views.TopBarView
 import fr.jaetan.jmedia.ui.SubScreen
 
 class SettingsView: SubScreen<SettingsViewModel>() {
     @Composable
+    override fun TopBar() {
+        TopBarView()
+    }
+
+    @Composable
     override fun Content() {
-        Text(R.string.settings.localized())
+        Box {
+            ContentView()
+        }
+    }
+
+    @Composable
+    override fun Dialogs() {
+        RemoveDataDialog()
     }
 }